+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/config/batch/StepConfig.java
@@ -1,0 +1,4 @@
+package com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.batch;
+
+public class StepConfig {
+}
