+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/config/quartz/QuartzJobConfig.java
@@ -28,8 +28,8 @@ import com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.coll
  */
 @Configuration
 public class SchedulerConfig {
-	public final String cronExpression;
-
+	private final String cronExpression;
+	private final String articleCollectorBatchJobDetailName = "rssBatchJob";
 	/**
 	 * 생성자 주입을 통해 Cron 표현식을 설정한다.
 	 *
@@ -43,19 +43,40 @@ public class SchedulerConfig {
 		this.cronExpression = cronExpression;
 	}
 
+	/**
+	 * RSS 수집 Quartz JobDetail 빈 등록.
+	 * Job 클래스는 {@link RssQuartzJob}이며 다음과 같은 설정을 포함한다:
+	 * - withIdentity("rssBatchJob"): Scheduler 내에서 이 Job을 고유하게 식별하기 위한 이름 지정
+	 * - storeDurably(): Trigger가 없더라도 Scheduler에 등록된 상태로 유지되도록 설정
+	 *
+	 * @return RSS 배치 작업용 JobDetail 객체
+	 * @author 함예정
+	 * @since 2025-05-10
+	 */
 	@Bean
-	public JobDetail rssBatchJobDetail() {
+	public JobDetail articleCollectorBatchJobDetail() {
 		return JobBuilder.newJob(RssQuartzJob.class)
-			.withIdentity("rssBatchJob")
+			.withIdentity("articleCollectorBatchJobDetail")
 			.storeDurably()
 			.build();
 	}
 
+	/**
+	 * RSS 수집 Quartz Trigger 빈 등록.
+	 * 다음과 같은 설정을 포함하며 {@link #articleCollectorBatchJobDetail()}에 연결된다:
+	 *  - forJob: 이 Trigger 가 어떤 Quartz Job 과 연관되어 실행될지를 지정
+	 * - withIdentity: Scheduler 내에서 이 Trigger 를 고유하게 식별하기 위한 이름 지정
+	 * - withSchedule: Cron 표현식을 사용하여 실행 주기 설정
+	 *
+	 * @return RSS 배치 작업용 Trigger 객체
+	 * @author 함예정
+	 * @since 2025-05-10
+	 */
 	@Bean
 	public Trigger rssBatchTrigger() {
 		return TriggerBuilder.newTrigger()
-			.forJob(rssBatchJobDetail())
-			.withIdentity("rssBatchTrigger")
+			.forJob(articleCollectorBatchJobDetail())
+			.withIdentity(articleCollectorBatchJobDetailName)
 			.withSchedule(CronScheduleBuilder.cronSchedule(cronExpression))
 			.build();
 	}
