+++ src/main/java/com/likelion/backendplus4/talkpick/batch/news/article/infrastructure/collector/config/quartz/QuartzTriggerConfig.java
@@ -1,0 +1,4 @@
+package com.likelion.backendplus4.talkpick.batch.news.article.infrastructure.collector.config.quartz;
+
+public class QuartzTriggerConfig {
+}
